requirements.txt
feedparser>=6.0.0
newspaper3k>=0.2.8
trafilatura>=1.6.0
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.18
gradio>=4.0.0
pydantic>=2.0.0
tqdm>=4.66.0
loguru>=0.7.0
openai>=1.0.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
python-dateutil>=2.8.2
listparser>=0.20
cfscrape
# 新增依赖
modelscope>=1.9.5
accelerate>=0.25.0
einops>=0.7.0
# pip install 'urllib3<2'
# pip install torch==2.5.1 torchvision==0.20.1 torchaudio==2.5.1 --index-url https://download.pytorch.org/whl/cu124
# CMAKE_ARGS="-DLLAVA_BUILD=OFF" pip install -U llama-cpp-python